Test.Hspec.MarkdownFormatter

Documentation

markdownFormatter :: String -> FilePath -> FormatConfig -> IO Format

Generates a markdowm-formatted test tree into a file.

Using the formatter requires to modify the configuration used by hspec to run the tests.

@@ import qualified MySpec as Spec

main :: IO () main = hspecWith defaultConfig{configFormat = Just (markdownFormatter "hspec-results.md")} MySpec.spec @@

The resulting markdown content follows these conventions:

The title is written as a level 1 header at the top of the file,

Every describe statement generates a new header one level below the enclosing one,

Every it or specify statement yields a list item,

Additional info output by a test (eg. QuickCheck's property count) will be added as pre-formatted text inside a foldable _Details_ section under its item.

NOTE: It seems the way to add formatters changed in 2.10 so perhaps this might not work in all settings.