Test.Hydra.Prelude

Documentation

createSystemTempDirectory :: String -> IO FilePath

Create a unique temporary directory.

failure :: (HasCallStack, MonadThrow m) => String -> m a

Fails a test with given error message. This function improves over existing expectationFailure by throwing a HUnitFailure exception containig the location of the error and providing better callstack context.

location :: HasCallStack => Maybe SrcLoc

Provides the source code location where this function is called. This relies on the CallStack information provided by GHC and to be useful requires all functions to be properly annotated.

failAfter :: (HasCallStack, MonadTimer m, MonadThrow m) => DiffTime -> m a -> m a

Fail some monadic action if it does not complete within given timeout. A DiffTime can be represented as a decimal number of seconds.

dualFormatter

Arguments

:: String
The name of the test suite run, for reporting purpose.

-> FormatConfig
Configuration, will be passed by the HSpec test runner.

-> IO Format

An HSpec test formatter that outputs both a JUnit formatted file and stdout test results.