Test.Network.Ports

Description

Utility module providing functions to find and allocate random ports.

Documentation

getRandomPort :: IO PortNumber

Find a TCPv4 port which is likely to be free for listening on localhost. This binds a socket, receives an OS-assigned port, then closes the socket.

Note that this is vulnerable to race conditions if another process binds the port returned by getRandomPort before this process does.

Do not use this unless you have no other option.

withFreePort :: (PortNumber -> IO a) -> IO a

Find a free TCPv4 port and pass it to the given action.

NOTE: Should be used only for testing, see getRandomPort for limitations.

randomUnusedTCPPorts :: Int -> IO [Int]

Find the specified number of free ports.

NOTE: Should be used only for testing, see getRandomPort for limitations.