Test.Plutus.Validator

Description

A helper module mostly wrapping the Alonzo.Tools' evaluateTransactionExecutionUnits with a much simpler API (just a plutus script).

This is generally handy to measure the execution of Plutus code outside of any context (e.g. an implementation of a data-structure on-chain or, as here, data encoders).

Documentation

wrapValidator :: (UnsafeFromData datum, UnsafeFromData redeemer) => (datum -> redeemer -> ScriptContext -> Bool) -> BuiltinData -> BuiltinData -> BuiltinData -> ()

Wrap a typed validator to get the basic Validator signature which can be passed to compile. Vendored from `plutus-ledger`. REVIEW: There might be better ways to name this than "wrap"

distanceExecutionUnits :: ExecutionUnits -> ExecutionUnits -> ExecutionUnits

evaluateScriptExecutionUnits :: ToData a => SerialisedScript -> a -> Either Text ExecutionUnits

pparams :: ProtocolParameters

Current (2023-04-12) mainchain parameters.

defaultMaxExecutionUnits :: ExecutionUnits

Max transaction execution unit budget of the current pparams.

transactionBodyFromScript :: ToScriptData a => SerialisedScript -> a -> (TxBody, UTxO)

Create an artifical transaction body which only spends the given script with given redeemer and a defaultDatum.

defaultDatum :: ()

The default datum used in transactionBodyFromScript.

data ExecutionUnits

Constructors

ExecutionUnits

Fields

executionSteps :: Natural

executionMemory :: Natural