Test.Hydra.Prelude

Documentation

createSystemTempDirectory :: String -> IO FilePath

Create a unique temporary directory.

failure :: (HasCallStack, MonadThrow m) => String -> m a

Fails a test with given error message. This function improves over existing expectationFailure by throwing a HUnitFailure exception containig the location of the error and providing better callstack context.

location :: HasCallStack => Maybe SrcLoc

Provides the source code location where this function is called. This relies on the CallStack information provided by GHC and to be useful requires all functions to be properly annotated.

failAfter :: (HasCallStack, MonadTimer m, MonadThrow m) => DiffTime -> m a -> m a

Fail some monadic action if it does not complete within given timeout. A DiffTime can be represented as a decimal number of seconds.

dualFormatter

Arguments

:: Text
The name of the test suite run, for reporting purpose.

-> FormatConfig
Configuration, will be passed by the HSpec test runner.

-> IO Format

An HSpec test formatter that outputs both a JUnit formatted file and stdout test results.

reasonablySized :: Gen a -> Gen a

Resize a generator to grow with the size parameter, but remains reasonably sized. That is handy when testing on data-structures that can be arbitrarily large and, when large entities don't really bring any value to the test itself.

It uses a square root function which makes the size parameter grows quadratically slower than normal. That is,

Normal Size ----------- 0 1 10 100 1000	Reasonable Size ---------------- 0 1 3 10 31

newtype ReasonablySized a

A QuickCheck modifier to make use of reasonablySized on existing types.

Constructors

ReasonablySized a

genericCoverTable :: forall a prop. (Show a, Enum a, Bounded a, Typeable a, Testable prop) => [a] -> prop -> Property

Like coverTable, but construct the weight requirements generically from the provided label.

data MyCoverLabel = Case1 | Case2 deriving (Enum, Bounded, Show)

forAll arbitrary $ a -> myProp a & genericCoverTable [a] & checkCoverage

===

forAll arbitrary $ a -> myProp a & coverTable MyCoverTable [(Case1, 50), (Case2, 50)] & tabulate MyCoverTable [head $ words $ show a] & checkCoverage

If myProp should take some data, use a product type around MyCoverLabel, for example:

type WithLabel a = (MyCoverLabel, a)

forAll arbitrary $ (label, input) -> myProp input & genericCoverTable [label] & checkCoverage

forAll2 :: (Testable property, Show a, Show b) => Gen a -> Gen b -> ((a, b) -> property) -> Property

Shorthand for using 2 generated values in a property.

HSpec re-exports

withTempDir :: String -> (FilePath -> IO r) -> IO r

Create a temporary directory for the given action to use. The directory is removed if and only if the action completes successfuly.

withLogFile :: FilePath -> (Handle -> IO a) -> IO a

Open given log file non-buffered in append mode and print a message with filepath to stderr on exceptions.

checkProcessHasNotDied :: Text -> ProcessHandle -> IO Void

Wait for process termination and do failure on non-zero exit code. This function is useful for end-to-end testing of external processes esp. in conjunction with race combinator:

@ withCreateProcess p $ _stdin _stdout _stderr processHandle -> do race_ (checkProcessHasNotDied "my-process" processHandle) doStuff @