Test.Plutus.Validator

Description

A helper module mostly wrapping the Alonzo.Tools' evaluateTransactionExecutionUnits with a much simpler API (just a plutus script).

This is generally handy to measure the execution of Plutus code outside of any context (e.g. an implementation of a data-structure on-chain or, as here, data encoders).

Documentation

wrapValidator :: (UnsafeFromData datum, UnsafeFromData redeemer) => (datum -> redeemer -> ScriptContext -> Bool) -> BuiltinData -> BuiltinData -> BuiltinData -> ()

Wrap a typed validator to get the basic Validator signature which can be passed to compile. Vendored from `plutus-ledger`. REVIEW: There might be better ways to name this than "wrap"

defaultMaxExecutionUnits :: ExUnits

Current (2022-04-01) mainchain parameters.

distanceExUnits :: ExUnits -> ExUnits -> ExUnits

evaluateScriptExecutionUnits :: ToData a => Validator -> a -> Either Text ExUnits

transactionFromScript :: ToData a => Validator -> a -> (ValidatedTx (BabbageEra StandardCrypto), UTxO (BabbageEra StandardCrypto))

newtype ExUnits

Constructors

WrapExUnits

Fields

unWrapExUnits :: ExUnits' Natural

Bundled Patterns

pattern ExUnits :: Natural -> Natural -> ExUnits